requirements.txt
beautifulsoup4
discord.py
fandom_py
requests
